Fix: Split Among Blueprints
